# frontend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Frametric API",
    "description": "Cinematic Analytics Platform — REST API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5168"
    }
  ],
  "paths": {
...
    "/api/v1/Analytics/dashboard": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Analytics/wrapped": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedSummaryDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Analytics/monthly-activity/{year}": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyActivityResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyActivityResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyActivityResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Analytics/top-directors": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DirectorLeaderboardDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DirectorLeaderboardDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DirectorLeaderboardDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Import/letterboxd": {
      "post": {
        "tags": [
          "Import"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Import/history": {
      "get": {
        "tags": [
          "Import"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportHistoryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportHistoryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportHistoryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Import/{id}": {
      "delete": {
        "tags": [
          "Import"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorCountDto": {
        "type": "object",
        "properties": {
          "actorName": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "averageRating": {
            "type": "number",
            "format": "double"
          },
          "watchedCount": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "profilePath": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "DashboardSummaryDto": {
        "required": [
          "totalWatchtimeMinutes",
          "totalWatches",
          "uniqueMoviesCount",
          "topGenres",
          "topDirectors",
          "topActors",
          "decadeBreakdown"
        ],
        "type": "object",
        "properties": {
          "totalWatchtimeMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "totalWatches": {
            "type": "integer",
            "format": "int32"
          },
          "uniqueMoviesCount": {
            "type": "integer",
            "format": "int32"
          },
          "topGenres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreCountDto"
            }
          },
          "topDirectors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectorCountDto"
            }
          },
          "topActors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorCountDto"
            }
          },
          "decadeBreakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecadeCountDto"
            }
          }
        }
      },
...
      "DecadeCountDto": {
        "required": [
          "decade",
          "count"
        ],
        "type": "object",
        "properties": {
          "decade": {
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "DirectorCountDto": {
        "type": "object",
        "properties": {
          "directorName": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "averageRating": {
            "type": "number",
            "format": "double"
          },
          "watchedCount": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "profilePath": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "DirectorLeaderboardDto": {
        "required": [
          "directorId",
          "name",
          "watchCount",
          "averageRating",
          "highestRatedMovieTitle"
        ],
        "type": "object",
        "properties": {
          "directorId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "averageRating": {
            "type": "number",
            "format": "double"
          },
          "highestRatedMovieTitle": {
            "type": "string"
          },
          "customAverageRating": {
            "type": "number",
            "format": "double",
            "default": null,
            "nullable": true
          }
        }
      },
...
      "GenreCountDto": {
        "type": "object",
        "properties": {
          "genreName": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "watchedCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ImportHistoryDto": {
        "required": [
          "id",
          "importDate",
          "rowCount",
          "status",
          "providerSource",
          "errorMessage"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "importDate": {
            "type": "string",
            "format": "date-time"
          },
          "rowCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "providerSource": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MonthlyActivityDto": {
        "required": [
          "month",
          "count"
        ],
        "type": "object",
        "properties": {
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MonthlyActivityResponseDto": {
        "required": [
          "monthlyActivity",
          "weeklyActivity"
        ],
        "type": "object",
        "properties": {
          "monthlyActivity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyWatchesDto"
            }
          },
          "weeklyActivity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeeklyWatchesDto"
            }
          }
        }
      },
...
      "MonthlyWatchesDto": {
        "required": [
          "month",
          "count"
        ],
        "type": "object",
        "properties": {
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "WeeklyWatchesDto": {
        "required": [
          "dayOfWeek",
          "count"
        ],
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "WrappedSummaryDto": {
        "required": [
          "year",
          "totalWatchtimeMinutes",
          "totalWatches",
          "uniqueMoviesCount",
          "topGenres",
          "topDirectors",
          "topActors",
          "decadeBreakdown",
          "monthlyActivity"
        ],
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "totalWatchtimeMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "totalWatches": {
            "type": "integer",
            "format": "int32"
          },
          "uniqueMoviesCount": {
            "type": "integer",
            "format": "int32"
          },
          "topGenres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreCountDto"
            }
          },
          "topDirectors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectorCountDto"
            }
          },
          "topActors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorCountDto"
            }
          },
          "decadeBreakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecadeCountDto"
            }
          },
          "monthlyActivity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyActivityDto"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste your JWT access token here (without the 'Bearer ' prefix).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}